+++ crates/sidecar/src/xserver/handlers/property/event.rs
@@ -182,7 +182,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                     if state.window_selects(source_window, EventMask::PROPERTY_CHANGE) {
                         state.pending_events.push(pn_event.clone());
                     }
-                    state.broadcast_event(source_window, u32::from(EventMask::PROPERTY_CHANGE), &pn_event);
+                    state.broadcast_event(source_window, EventMask::PROPERTY_CHANGE, &pn_event);
                 }
 
                 // Determine the new WM state and broadcast to frontend
@@ -255,7 +255,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                         {
                             state.pending_events.push(cn.clone());
                         }
-                        state.broadcast_event(source_window, u32::from(EventMask::STRUCTURE_NOTIFY), &cn);
+                        state.broadcast_event(source_window, EventMask::STRUCTURE_NOTIFY, &cn);
 
                         // Send Expose so the client redraws at the new size
                         if state.window_selects(source_window, EventMask::EXPOSURE)
@@ -341,7 +341,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                             {
                                 state.pending_events.push(cn.clone());
                             }
-                            state.broadcast_event(source_window, u32::from(EventMask::STRUCTURE_NOTIFY), &cn);
+                            state.broadcast_event(source_window, EventMask::STRUCTURE_NOTIFY, &cn);
 
                             // Expose for redraw
                             if state.window_selects(source_window, EventMask::EXPOSURE)
@@ -510,7 +510,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                     if state.window_selects(source_window, EventMask::PROPERTY_CHANGE) {
                         state.pending_events.push(pn_event.clone());
                     }
-                    state.broadcast_event(source_window, u32::from(EventMask::PROPERTY_CHANGE), &pn_event);
+                    state.broadcast_event(source_window, EventMask::PROPERTY_CHANGE, &pn_event);
                 }
                 return Vec::new();
             }
@@ -657,7 +657,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                         if state.window_selects(source_window, EventMask::PROPERTY_CHANGE) {
                             state.pending_events.push(pn.clone());
                         }
-                        state.broadcast_event(source_window, u32::from(EventMask::PROPERTY_CHANGE), &pn);
+                        state.broadcast_event(source_window, EventMask::PROPERTY_CHANGE, &pn);
                     }
                     if let Some(uuid) = state.window_uuid(source_window) {
                         let _ = state.update_tx.send((
@@ -697,7 +697,7 @@ pub(crate) fn handle_send_event(state: &mut ClientState, req: &SendEventRequest)
                         if state.window_selects(source_window, EventMask::PROPERTY_CHANGE) {
                             state.pending_events.push(pn.clone());
                         }
-                        state.broadcast_event(source_window, u32::from(EventMask::PROPERTY_CHANGE), &pn);
+                        state.broadcast_event(source_window, EventMask::PROPERTY_CHANGE, &pn);
                     }
                     if let Some(uuid) = state.window_uuid(source_window) {
                         let _ = state.update_tx.send((
